requirements.txt
# python v 3.12.4
numpy==1.26.4
pandas==2.2.3
protobuf==4.21.6
typing-extensions==4.12.1
python-dotenv==1.0.0
fastapi==0.115.0
openai==1.54.3
qdrant-client==1.9.1
coloredlogs==15.0.1
torch==2.2.2
transformers==4.41.1
accelerate==0.30.1
pydantic-settings==2.2.1
beautifulsoup4==4.12.3
scikit-learn==1.5.2
asyncpg==0.29.0
uvicorn==0.31.0
pydantic==2.9.2
fastapi-limiter==0.1.6
slowapi==0.1.9
aiohttp==3.10.9
python-multipart==0.0.12
asyncpg==0.29.0
greenlet==3.1.1
